+++ CHANGES.md
@@ -1,5 +1,6 @@
 # FuzzyFileNav 2.0.0
 
+- **NEW**: Allow opening folder in new window.
 - **NEW**: Commands are available from the command palette.
 - **NEW**: Move context menu items to tab context menu as commands don't directly operate on view content.
 - **FIX**: Don't copy `.` and `..`.